requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
alabaster==0.7.12
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
appdirs==1.4.4
argh==0.26.2
asgiref==3.2.10
asn1crypto==1.3.0
astor==0.8.1
astroid==2.4.2
astropy==4.0.1.post1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
auto-sklearn==0.7.0
autocorrect==2.0.0
autokeras==1.0.2
automl==2.9.9
autopep8==1.5.3
av==8.0.2
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.0.1
beautifulsoup4==4.9.1
BigDL==0.10.0
bitarray==1.2.2
bkcharts==0.2
bleach==3.1.5
blinker==1.4
blis==0.4.1
bokeh==2.1.0
boto==2.49.0
boto3==1.14.3
botocore==1.17.3
Bottleneck==1.3.2
cachetools==4.1.0
catalogue==1.0.0
catboost==0.23.2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
clyent==1.2.2
cmdstanpy==0.9.5
colorama==0.4.3
conda==4.8.3
conda-build==3.18.11
conda-pack==0.4.0
conda-package-handling==1.6.0
conda-verify==3.4.2
ConfigSpace==0.4.13
contextlib2==0.6.0.post1
convertdate==2.2.1
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.17
cytoolz==0.10.1
dask==2.18.1
dask-glm==0.2.0
dask-ml==1.5.0
dask-xgboost==0.1.10
deap==1.3.1
decorator==4.4.2
decord==0.3.9
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.2
distributed==2.18.0
Django==3.1.1
dlib==19.20.0
docutils==0.15.2
download==0.3.5
easydict==1.9
entrypoints==0.3
enum-compat==0.0.3
ephem==3.7.7.1
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
facemask==0.1.1
fastcache==1.1.0
fbprophet==0.6
featuretools==0.16.0
filelock==3.0.12
firebase==3.0.1
flake8==3.8.3
Flask==1.1.2
fsspec==0.7.4
future==0.18.2
fvcore==0.1.1.post20200616
gast==0.3.3
gdown==3.11.1
gevent==20.6.1
glob2==0.7
gluoncv==0.7.0
gluonts==0.5.0
gmpy2==2.0.8
google==2.0.3
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googletrans==3.0.0
graphviz==0.14
greenlet==0.4.16
grpcio==1.29.0
h11==0.9.0
h2==3.2.0
h2o==3.30.0.4
h5py==2.10.0
HeapDict==1.0.1
helpdev==0.7.1
holidays==0.9.12
hpack==3.0.0
hstspreload==2020.6.16
httpcore==0.9.1
httpx==0.13.3
hummingbird-ml==0.0.2
hyperframe==5.2.0
hypothesis==5.16.1
idna==2.9
imagecodecs==2020.5.30
imageio==2.8.0
imageio-ffmpeg==0.4.2
imagesize==1.2.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==1.6.1
imutils==0.5.3
install==1.3.3
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.3
jellyfish==0.8.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.0.1
keyring==21.2.1
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
langdetect==1.0.8
language-check==1.1
lazy-import==0.2.2
lazy-object-proxy==1.4.3
liac-arff==2.4.0
libarchive-c==2.9
libra==1.0.3
lief==0.10.1
lightgbm==2.3.1
llvmlite==0.31.0
locket==0.2.0
lockfile==0.12.2
LunarCalendar==0.0.9
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
moviepy==1.0.3
mpmath==1.1.0
msgpack==1.0.0
multidict==4.7.6
multipledispatch==0.6.0
multiprocess==0.70.10
murmurhash==1.0.2
mxnet==1.6.0
mxnet-cu101mkl==1.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
nose==1.3.7
notebook==6.0.3
numba==0.49.0
numexpr==2.7.1
numpy==1.18.5
numpydoc==1.0.0
oauthlib==3.1.0
olefile==0.46
onnx==1.7.0
onnxconverter-common==1.7.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandas-bokeh==0.5
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==14.0.1
pathlib2==2.3.5
pathos==0.2.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip-upgrade-outdated==1.5
pkginfo==1.5.0.1
plac==1.1.3
plotly==4.8.1
pluggy==0.13.1
ply==3.11
pmdarima==1.6.1
pooch==1.1.1
portalocker==1.7.0
pox==0.2.8
ppft==1.6.6.1
pprint==0.1
preshed==3.0.2
prince==0.7.0
proglog==0.1.9
progressbar2==3.51.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pure-predict==0.0.4
py==1.8.2
py-spy==0.3.3
py4j==0.10.9
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycocotools==2.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==1.5.1
pydeck==0.4.0b2
pydocstyle==5.0.2
pydot==1.4.1
pydub==0.24.1
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
PyMeeus==0.3.7
pynisher==0.5.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyrfr==0.8.0
pyrsistent==0.16.0
PySocks==1.7.1
pyspark==3.0.0
pystan==2.19.1.1
pytest==5.4.3
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.7.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-firebase==1.2
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python-Levenshtein==0.12.0
python-utils==2.4.0
PyTurboJPEG==1.4.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
ray==0.8.5
redis==3.5.3
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.4.0
rope==0.17.0
rsa==4.6
Rtree==0.9.4
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
SDDect==0.1.0
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
sentencepiece==0.1.91
setuptools-git==1.2
sightengine==1.4.0
simplegeneric==0.8.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.15.0
sk-dist==0.1.9
sklearn==0.0
sklearn-deap2==0.2.2
# Editable install with no version control (slowfast==1.0)
-e /home/hack_abhilash/anaconda3/lib/python3.7/site-packages
smac==0.12.2
sniffio==1.1.0
snowballstemmer==2.0.0
sobol-seq==0.2.0
SocialDistance==0.1.1
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
spacy-langdetect==0.1.2
SpeechRecognition==3.8.1
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.2
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
srsly==1.0.2
statsmodels==0.11.1
stopit==1.1.2
streamlit==0.63.1
streamlit-theme==0.58.0
sympy==1.6
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.6.3
tokenizers==0.7.0
toml==0.10.1
toolz==0.10.0
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
TPOT==0.11.5
tqdm==4.46.1
traitlets==4.3.3
transformers==2.11.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
ujson==3.0.0
unicodecsv==0.14.1
update-checker==0.17
urllib3==1.25.9
utils==1.0.1
validators==0.16.0
vega-datasets==0.8.0
wasabi==0.7.1
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
xgboost==1.1.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
xmltodict==0.12.0
yacs==0.1.7
yapf==0.30.0
yarl==1.4.2
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
